Encapsulate map in its own component
